+++ CHANGES.rst
@@ -10,11 +10,11 @@ TidyPy Change Log
 
 **Enhancements**
 
+* Added ``manifest`` and ``pydiatra`` tools.
 * Upgraded the ``pylint`` tool.
 * Upgraded the ``pep8-naming`` plugin of the ``pycodestyle`` tool.
 * Added some convenience handling of the ``License`` vs ``Licence`` and
   ``LicenceClassifier`` vs ``LicenseClassifier`` codes reported by ``pyroma``.
-* Added a ``manifest`` tool.
 
 **Fixes**